get min max price API for front end usage

+++ src/car/car.entity.ts
@@ -36,7 +36,7 @@ export class Car {
   @Column()
   offerType: string;
 
-  @Column('float')
+  @Column('int')
   price: number;
 
   @Column('int')
